pyproject.toml
[project]
name = "pavepo-test"
version = "0.1.0"
readme = "README.md"
repository = "https://github.com/SuperLalka/pavepo-test.git"

[tool.poetry]
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
uvicorn = {extras = ["standard"], version = "0.34.0"}
fastapi = "0.115.12"
psycopg2-binary = "2.9.10"
asyncpg = "0.30.0"
SQLAlchemy = "2.0.39"
sqlalchemy-file = "0.6.0"
alembic = "1.15.1"
python-jose = {extras = ["cryptography"], version = "3.4.0"}
passlib = "1.7.4"
fastapi-oauth2 = "1.3.0"
fasteners = "0.19"
aiofiles = "24.1.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
